interface faltantes
